Moved CompactionExecutor classes into a separate new package.
