+++ ql/src/java/org/apache/hadoop/hive/ql/txn/compactor/Worker.java
@@ -29,6 +29,8 @@ import org.apache.hadoop.hive.metastore.api.Table;
 import org.apache.hadoop.hive.metastore.conf.MetastoreConf;
 import org.apache.hadoop.hive.metastore.metrics.MetricsConstants;
 import org.apache.hadoop.hive.ql.log.PerfLogger;
+import org.apache.hadoop.hive.ql.txn.compactor.service.CompactionExecutor;
+import org.apache.hadoop.hive.ql.txn.compactor.service.CompactionExecutorFactory;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.thrift.TException;
 import org.slf4j.Logger;
